[skip ci]
• emcomposition2.py:
  - cast return value of argmin to double()
  - _construct_combined_scores_node: hack that adds dimensions to return value of _gen_pytorch_fct() to fix torch_stacking (currently commented out)

+++ psyneulink/library/compositions/emcomposition/emcomposition2.py
@@ -1230,7 +1230,9 @@ class EMComposition2(AutodiffComposition):
             softmax_func = softmax_function._gen_pytorch_fct(device, local_context)
             def func(variable):
                 variable = variable.squeeze()
-                return [softmax_func(variable[0]), torch.atleast_1d(torch.argmin(variable[1]))]
+                # EM2 BREADCRUMB: SHOULD REPLACE CASTING TO DOUBLE BELOW WITH CASTING TO DEFAULT PRECISION
+                return [softmax_func(variable[0]), torch.atleast_1d(torch.argmin(variable[1]).double())]
+                # return [[[softmax_func(variable[0]), torch.atleast_1d(torch.argmin(variable[1]).double())]]]
             return func
 
         combined_scores_function = UserDefinedFunction(_combined_scores_function,
